Code for part 2
Part 2 is only halfway completed. A class named GeneralPractitioner was created in the child package, and a new variable named 'speicalty' was defined and assigned the value of 'general practitioner', along with an output statement. At the same time, this class inherits three variables of HealthProfessional from the base package, which may require