New testing code（For part2）
The last test was to create an object for HealthProfessioanl with only three variable values. And this time, the test code was written for part 2, creating an object for GeneralPractitioner and Dentist respectively, and outputting it. There are four variables in this test. At the same time, I added two slashes (which became comments) to the object and output statement code created last time instead of deleting them. This way, the teacher can run these statements at any time, so they just became comments