Correct the code
Although I was able to output it last time, one variable output result was null. Then I noticed that the first letter of a word was accidentally written in uppercase. I changed the uppercase S to lowercase s, and the result was no longer null. The testing for Part 2 has been completely completed