Create 5 objects (Part 3)
I created 3 general practitioner objects and 2 dentist objects according to the requirements of Part 3, and then wrote the code to output the statements (in the main method). At the same time, I separated the previous test output code and annotated it with Part 1 and Part 2 (Part 1 and Part 2 that meet the job requirements). (At the same time, the creation objects and output statement codes of Part 1 and Part 2 were all marked with double slashes, becoming comments. If you want to test the objects and outputs created by Part 1 and Part 2 in A1, simply remove the double slashes.)