Finish part4
The code for this section has been completed, and I have defined some appointment variables in the Appointment class, such as the patient's name, phone number, and appointment time. At the same time, it also inherits the variable information of doctors in the HealthProfessional class. For testing purposes, I created an appointment object and a doctor object (here I selected a general practitioner) in the main method of this class, and then printed them out (this time the code only has one appointment information)