Part 1 : No testing objects
I created a base package in my assignment and created a class inside it: HealthProfessinal， Then define three private type variables, assign them with a constructor function, initialize the three variables by constructor function which have parameter ("this"), and output them using printInfo. Meanwhile, the three variables are of private type, so it is necessary to use Getter to obtain the values of the variables and then use Setter to modify it. This is the code for Part 1. But no objects have been created and tested yet