part 2 General doctor
In part2， Only a part has been completed, creating the child folder and creating the GeneralPractitioner class in it, defining a variable (private): specialty (without Getter and Setter methods), and then initializing four variables with a parameterized constructor. ID, name, basicInfo use the super method, and specialty use "this". ID, name, and basicInfo come from the HealthProfessional class, so we need to import the base.HealthProfessional package and  class. At the same time, we also need to let the GeneralPractitioner class inherit (extends) HealthProfessional, so that all four variables can be used. Finally, rewrite the PrintInfo method (from the parent class), and the output method can also be used. (At this time, there are still no objects in the HealthProfessional class)