Test 1 appointment
In the Appointment class, I only imported 2 packages, import base.HealthProfessional； import child.GeneralPractitioner;  After all, make appointments with a general practitioner for patients. I have defined three variables representing patient information in this class, as well as a HealthProfessional type doctor variable representing the doctor's data. Then use the constructor to assign values, and initialize these variables with a parameterized constructor (all using 'this'). Then I used the if statement in the printAppointDetails function to print the patients' personal information first. If there is no doctor, there is no appointment. But I created an object for a general practitioner in the main method, so printInfo in if will be called, and then I can output the doctor information. Additionally, I created an appointment object in the main method, so there won't be no doctors left. Then, I use printAppointmentDetails to output the results