Part 5 is OK
Created the last class: Main， And I imported 3 packages (3 classes): import base.HealthProfessional;  import child.Dentist;  import child.GeneralPractitioner;  In addition, I also imported packages of arrays: import java.util.ArrayList; In order to ensure that users can input data, a package for inputting data has been imported: import java.util.Scanner; 
Firstly, I defined an array: appointments， This is used to store historical records, and then I first created two existing appointments (one for a dentist and one for a general practitioner) using the createAppointment method. In the createAppointment function, I used an if statement to determine if the four questions that need to be answered for the appointment are missing. If they are missing, the appointment cannot be created, but if they are not, the new element is added to the array. Then use a for each loop in the printExistingAppointments function to iterate through the array, and use printAppointDetails in the loop to output twice (2 elements) (I wrote an if statement, if the array is empty, there is no history reservation), and then you can use printExistingAppointments to output the element information of the array. I first created two existing appointments (one dentist and one general practitioner), then used a for each loop in the printExistingAppointments function to iterate through the array. In the loop, I used printAppointmentDetails to output twice (2 elements) (I wrote an if statement, if the array is empty, there are no historical appointments), and then I can use printExistingAppointments to output the element information of the array
Next, type in the data to create an appointment, answer questions such as name, and use the if statement to select a doctor. 1 is a general practitioner, and 2 is a dentist. To create a reservation, simply answer 4 questions. A new element will take in the array ( appointment ) and output it by .printExistingAppointments Of course In the cancelBooking method,Alternatively, you can enter your phone number to cancel the reservation. Use a for loop to iterate through the array to retrieve all elements (reservation records) and obtain all phone numbers. Then, I use an if function to check if the input phone number matches the phone number in the database. If it matches, the cancellation is successful. This element(appointment) will remove from array , If it does not match, the cancellation fails